+++ generator/src/main/java/com/linkedin/pegasus/generator/FileFormatDataSchemaParser.java
@@ -49,16 +49,21 @@ import java.util.jar.JarFile;
 public class FileFormatDataSchemaParser
 {
   static final String SCHEMA_PATH_PREFIX = SchemaDirectoryName.PEGASUS.getName() + "/";
-  static final String EXTENSION_PATH_ENTRY = SchemaDirectoryName.EXTENSIONS.getName() + "/";
-  private final String _resolverPath;
   private final DataSchemaResolver _schemaResolver;
   private final DataSchemaParserFactory _schemaParserFactory;
+  private final List<SchemaDirectoryName> _sourceDirectories;
 
   public FileFormatDataSchemaParser(String resolverPath, DataSchemaResolver schemaResolver, DataSchemaParserFactory schemaParserFactory)
   {
-    _resolverPath = resolverPath;
+    this(schemaResolver, schemaParserFactory, schemaResolver.getSchemaDirectories());
+  }
+
+  public FileFormatDataSchemaParser(DataSchemaResolver schemaResolver,
+      DataSchemaParserFactory schemaParserFactory, List<SchemaDirectoryName> sourceDirectories)
+  {
     _schemaResolver = schemaResolver;
     _schemaParserFactory = schemaParserFactory;
+    _sourceDirectories = sourceDirectories;
   }
 
   public DataSchemaParser.ParseResult parseSources(String[] sources) throws IOException
@@ -98,7 +103,7 @@ public class FileFormatDataSchemaParser
                 final JarEntry entry = entries.nextElement();
                 if (!entry.isDirectory() &&
                     entry.getName().endsWith(_schemaParserFactory.getLanguageExtension()) &&
-                    (entry.getName().startsWith(_schemaResolver.getSchemasDirectoryName().getName() + "/")))
+                    shouldParseFile(entry.getName()))
                 {
                   parseJarEntry(jarFile, entry, result);
                   result.getSourceFiles().add(sourceFile);
@@ -150,6 +155,17 @@ public class FileFormatDataSchemaParser
     }
   }
 
+  private boolean shouldParseFile(String path)
+  {
+    for (SchemaDirectoryName schemaDirectoryName : _sourceDirectories)
+    {
+      if (schemaDirectoryName.matchesJarFilePath(path))
+      {
+        return true;
+      }
+    }
+    return false;
+  }
   /**
    * Parse a source that specifies a file (not a fully qualified schema name).
    *