+++ generator/src/test/java/com/linkedin/pegasus/generator/TestDataSchemaParser.java
@@ -19,7 +19,7 @@ package com.linkedin.pegasus.generator;
 import com.linkedin.data.schema.DataSchema;
 import com.linkedin.data.schema.DataSchemaLocation;
 import com.linkedin.data.schema.NamedDataSchema;
-import com.linkedin.data.schema.resolver.ExtensionsDataSchemaResolver;
+import com.linkedin.data.schema.resolver.SchemaDirectoryName;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -28,6 +28,7 @@ import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -47,7 +48,7 @@ public class TestDataSchemaParser
 {
   private static final String FS = File.separator;
   private static final String testDir = System.getProperty("testDir", new File("src/test").getAbsolutePath());
-  private static final String pegasusDir = testDir + FS + "resources" + FS + "generator";
+  private static final String TEST_RESOURCES_DIR = testDir + FS + "resources" + FS + "generator";
 
   private File _tempDir;
   private File _dataTemplateTargetDir;
@@ -82,20 +83,82 @@ public class TestDataSchemaParser
     String tempDirectoryPath = _tempDir.getAbsolutePath();
     String jarFile = tempDirectoryPath + FS + "test.jar";
 
-    String pegasusFile = pegasusDir + FS + pegasusFilename;
-    String pegasusFileInJar = SCHEMA_PATH_PREFIX + pegasusFile;
+    String pegasusFile = TEST_RESOURCES_DIR + FS + pegasusFilename;
+    String pegasusFileInJar = SCHEMA_PATH_PREFIX + pegasusFilename;
     createTempJarFile(Collections.singletonMap(pegasusFile, pegasusFileInJar), jarFile);
 
-    DataSchemaParser parser = new DataSchemaParser(tempDirectoryPath);
+    DataSchemaParser parser = new DataSchemaParser.Builder(tempDirectoryPath).build();
     DataSchemaParser.ParseResult parseResult = parser.parseSources(new String[]{jarFile});
     // Two schemas, WithoutResolverExample and InlineRecord (defined inline in WithoutResolverExample)
-    assertEquals(parseResult.getSchemaAndLocations().size(), 2);
+    assertEquals(parseResult.getSchemaAndLocations().size(), expectedSchemas.length);
     Set<String> schemaNames = parseResult.getSchemaAndLocations().keySet().stream().map(DataSchema::getUnionMemberKey).collect(
         Collectors.toSet());
     for (String schema : expectedSchemas)
     {
       assertTrue(schemaNames.contains(schema));
     }
+    parseResult.getSchemaAndLocations().values().forEach(loc -> assertEquals(loc.getSourceFile().getAbsolutePath(), jarFile));
+  }
+
+  @Test(dataProvider = "inputFiles")
+  public void testCustomSourceSchemaDirectory(String pegasusFilename, String[] expectedSchemas) throws Exception
+  {
+    String tempDirectoryPath = _tempDir.getAbsolutePath();
+    String jarFile = tempDirectoryPath + FS + "test.jar";
+    SchemaDirectoryName customSchemaDirectory = () -> "custom";
+    String pegasusFile = TEST_RESOURCES_DIR + FS + pegasusFilename;
+    String pegasusFileInJar = customSchemaDirectory.getName() + "/" + pegasusFilename;
+    createTempJarFile(Collections.singletonMap(pegasusFile, pegasusFileInJar), jarFile);
+
+    // Load with default parser, this will return zero scheams.
+    DataSchemaParser parser = new DataSchemaParser.Builder(tempDirectoryPath).build();
+    DataSchemaParser.ParseResult parseResult = parser.parseSources(new String[]{jarFile});
+    assertEquals(parseResult.getSchemaAndLocations().size(), 0);
+
+    // Now create a parser with custom directory as source
+    parser = new DataSchemaParser.Builder(tempDirectoryPath)
+        .setSourceDirectories(Collections.singletonList(customSchemaDirectory))
+        .build();
+    parseResult = parser.parseSources(new String[]{jarFile});
+    assertEquals(parseResult.getSchemaAndLocations().size(), expectedSchemas.length);
+    Set<String> schemaNames = parseResult.getSchemaAndLocations().keySet().stream().map(DataSchema::getUnionMemberKey).collect(
+        Collectors.toSet());
+    for (String schema : expectedSchemas)
+    {
+      assertTrue(schemaNames.contains(schema));
+    }
+    parseResult.getSchemaAndLocations().values().forEach(loc -> assertEquals(loc.getSourceFile().getAbsolutePath(), jarFile));
+  }
+
+  @Test
+  public void testCustomResolverSchemaDirectory() throws Exception
+  {
+    String tempDirectoryPath = _tempDir.getAbsolutePath();
+    String jarFile = tempDirectoryPath + FS + "test.jar";
+    String schemaDir = TEST_RESOURCES_DIR + FS + "extensionSchemas";
+    SchemaDirectoryName customSchemaDirectory = () -> "custom";
+    Map<String, String> entryToFileMap = new HashMap<>();
+    // FooExtensions is in "extensions" directory and references "Foo" from "custom" directory.
+    entryToFileMap.put(schemaDir + FS + "pegasus/Foo.pdl", "custom/Foo.pdl");
+    entryToFileMap.put(schemaDir + FS + "extensions/FooExtensions.pdl", "extensions/FooExtensions.pdl");
+    createTempJarFile(entryToFileMap, jarFile);
+
+    List<SchemaDirectoryName> resolverDirectories = Arrays.asList(
+        SchemaDirectoryName.EXTENSIONS, customSchemaDirectory);
+    List<SchemaDirectoryName> sourceDirectories = Collections.singletonList(SchemaDirectoryName.EXTENSIONS);
+    DataSchemaParser parser = new DataSchemaParser.Builder(jarFile)
+        .setResolverDirectories(resolverDirectories)
+        .setSourceDirectories(sourceDirectories)
+        .build();
+
+    DataSchemaParser.ParseResult parseResult = parser.parseSources(new String[]{jarFile});
+    parseResult = parser.parseSources(new String[]{jarFile});
+    // Foo and FooExtensions
+    assertEquals(parseResult.getSchemaAndLocations().size(), 2);
+    Set<String> schemaNames = parseResult.getSchemaAndLocations().keySet().stream().map(DataSchema::getUnionMemberKey).collect(
+        Collectors.toSet());
+    assertTrue(schemaNames.contains("FooExtensions"));
+    assertTrue(schemaNames.contains("Foo"));
     parseResult.getSchemaAndLocations().values().forEach(loc -> assertEquals(loc.getSourceFile().getAbsolutePath(), jarFile));
   }
 
@@ -150,14 +213,19 @@ public class TestDataSchemaParser
   {
     String tempDirectoryPath = _tempDir.getAbsolutePath();
     String jarFile = tempDirectoryPath + FS + "test.jar";
-    String schemaDir = pegasusDir + FS + "extensionSchemas";
+    String schemaDir = TEST_RESOURCES_DIR + FS + "extensionSchemas";
     Map<String, String> entryToFileMap = Arrays.stream(files).collect(Collectors.toMap(
         filename -> schemaDir + FS + filename,
         filename -> filename));
     createTempJarFile(entryToFileMap, jarFile);
 
-    ExtensionsDataSchemaResolver resolver = new ExtensionsDataSchemaResolver(jarFile);
-    DataSchemaParser parser = new DataSchemaParser(jarFile, resolver);
+    List<SchemaDirectoryName> resolverDirectories = Arrays.asList(
+        SchemaDirectoryName.EXTENSIONS, SchemaDirectoryName.PEGASUS);
+    List<SchemaDirectoryName> sourceDirectories = Collections.singletonList(SchemaDirectoryName.EXTENSIONS);
+    DataSchemaParser parser = new DataSchemaParser.Builder(jarFile)
+        .setResolverDirectories(resolverDirectories)
+        .setSourceDirectories(sourceDirectories)
+        .build();
     DataSchemaParser.ParseResult parseResult = parser.parseSources(new String[]{jarFile});
     Map<DataSchema, DataSchemaLocation> extensions = parseResult.getExtensionDataSchemaAndLocations();
     assertEquals(extensions.size(), expectedExtensions.length);
@@ -174,13 +242,18 @@ public class TestDataSchemaParser
   @Test(dataProvider = "entityRelationshipInputFiles")
   public void testSchemaFilesInExtensionPathInFolder(String[] files, String[] expectedExtensions) throws Exception
   {
-    String pegasusWithFS = pegasusDir + FS;
+    String pegasusWithFS = TEST_RESOURCES_DIR + FS;
     String resolverPath = pegasusWithFS + "extensionSchemas/extensions:"
         + pegasusWithFS + "extensionSchemas/others:"
         + pegasusWithFS + "extensionSchemas/pegasus";
-    ExtensionsDataSchemaResolver resolver = new ExtensionsDataSchemaResolver(resolverPath);
-    DataSchemaParser parser = new DataSchemaParser(resolverPath, resolver);
-    String[] schemaFiles = Arrays.stream(files).map(casename -> pegasusDir + FS + "extensionSchemas" + FS + casename).toArray(String[]::new);
+    List<SchemaDirectoryName> resolverDirectories = Arrays.asList(
+        SchemaDirectoryName.EXTENSIONS, SchemaDirectoryName.PEGASUS);
+    List<SchemaDirectoryName> sourceDirectories = Collections.singletonList(SchemaDirectoryName.EXTENSIONS);
+    DataSchemaParser parser = new DataSchemaParser.Builder(resolverPath)
+        .setResolverDirectories(resolverDirectories)
+        .setSourceDirectories(sourceDirectories)
+        .build();
+    String[] schemaFiles = Arrays.stream(files).map(casename -> TEST_RESOURCES_DIR + FS + "extensionSchemas" + FS + casename).toArray(String[]::new);
     DataSchemaParser.ParseResult parseResult = parser.parseSources(schemaFiles);
     Map<DataSchema, DataSchemaLocation> extensions = parseResult.getExtensionDataSchemaAndLocations();
     assertEquals(extensions.size(), expectedExtensions.length);
@@ -221,12 +294,12 @@ public class TestDataSchemaParser
   @Test(dataProvider = "ERFilesForBaseSchema")
   public void testParseResultToGetBaseSchemas(String[] files, String[] expectedSchemaNames) throws Exception
   {
-    String pegasusWithFS = pegasusDir + FS;
+    String pegasusWithFS = TEST_RESOURCES_DIR + FS;
     String resolverPath = pegasusWithFS + "extensionSchemas/extensions:"
         + pegasusWithFS + "extensionSchemas/others:"
         + pegasusWithFS + "extensionSchemas/pegasus";
-    DataSchemaParser parser = new DataSchemaParser(resolverPath);
-    String[] schemaFiles = Arrays.stream(files).map(casename -> pegasusDir + FS + "extensionSchemas" + FS + casename).toArray(String[]::new);
+    DataSchemaParser parser = new DataSchemaParser.Builder(resolverPath).build();
+    String[] schemaFiles = Arrays.stream(files).map(casename -> TEST_RESOURCES_DIR + FS + "extensionSchemas" + FS + casename).toArray(String[]::new);
     DataSchemaParser.ParseResult parseResult = parser.parseSources(schemaFiles);
     Map<DataSchema, DataSchemaLocation> bases = parseResult.getBaseDataSchemaAndLocations();
     assertEquals(bases.size(), expectedSchemaNames.length);
@@ -248,18 +321,18 @@ public class TestDataSchemaParser
     Map<String, String> jarFiles = new HashMap<>();
 
     // Add the source PDL file to the pegasus directory
-    String pdlFile = pegasusDir + FS + "WithoutResolverExamplePdl.pdl";
+    String pdlFile = TEST_RESOURCES_DIR + FS + "WithoutResolverExamplePdl.pdl";
     String pdlJarDestination = SCHEMA_PATH_PREFIX + "WithoutResolverExamplePdl.pdl";
     jarFiles.put(pdlFile, pdlJarDestination);
 
     // Translated PDSC files go to "legacyPegasusSchemas", which should be ignored by parser.
-    String translatedPegasusFile = pegasusDir + FS + "WithoutResolverExample.pdsc";
+    String translatedPegasusFile = TEST_RESOURCES_DIR + FS + "WithoutResolverExample.pdsc";
     String translatedFileDestination = "legacyPegasusSchemas/WithoutResolverExample.pdsc";
 
     jarFiles.put(translatedPegasusFile, translatedFileDestination);
     createTempJarFile(jarFiles, jarFile);
 
-    DataSchemaParser parser = new DataSchemaParser(tempDirectoryPath);
+    DataSchemaParser parser = new DataSchemaParser.Builder(tempDirectoryPath).build();
     DataSchemaParser.ParseResult parseResult = parser.parseSources(new String[]{jarFile});
     // Two schemas, WithoutResolverExample and InlineRecord (defined inline in WithoutResolverExample)
     assertEquals(parseResult.getSchemaAndLocations().size(), 2);